Add files via upload
In this project, I have made interesting dashboards using different visualisations tools on the country_clean dataset. There is one dashboard in which you can select the latitude and longitude point and it will show the place on the world map. The screen captures of all the dashboards are present in the repository